Add get player data method to IGameServiceClient
If supported by the Game Service, this method is used to
invoke a call to fetch player related data.

Implement method stubs to submodules which just returns false when called.

Add IPlayerData interface which describes the PlayerData data model.

Add IPlayerDataResponseListener interface which declares a callback method
when player data is received from the Game Service.